feat: Lv5. 예외처리 추가
- 일정 추가에서 등록된 작성자명과 비밀번호가 아닌 경우 예외처리
- 일정 단건 조회시 해당 일정이 존재하지 않는 경우 예외처리
- 일정 업데이트에서 등록된 작성자명과 비밀번호가 아닌 경우 예외처리
- 일정 수정에서 일정 작성자가 아닌 사람이 수정할 경우 예외처리
- 일정 수정이 존재하지 않는 경우 예외처리
- 일정 삭제에서 등록된 작성자명과 비밀번호가 아닌 경우 예외처리
- 일정 삭제에서 일정 작성자가 아닌 사람이 삭제할 경우 예외처리

+++ src/main/java/com/example/calendar/service/CalendarServiceImpl.java
@@ -10,8 +10,10 @@ import com.example.calendar.repository.CalendarRepository;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.List;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 public class CalendarServiceImpl implements CalendarService {
@@ -25,12 +27,17 @@ public class CalendarServiceImpl implements CalendarService {
     @Override
     public CalendarResponseDto createSchedule(CalendarRequestDto dto) {
 
-        Writer writer = new Writer(dto.getWriterName(), dto.getPassword());
+        Long writerId = calendarRepository.findWriterIdByNameAndPassword(dto.getWriterName(), dto.getPassword());
 
+        if (writerId == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "작성자명 혹은 비밀번호가 틀렸습니다.");
+        }
+
+        Writer writer = new Writer(dto.getWriterName(), dto.getPassword());
         Calendar calendar = new Calendar(dto.getTodo(), Timestamp.valueOf(LocalDateTime.now()).toLocalDateTime(),
                 Timestamp.valueOf(LocalDateTime.now()).toLocalDateTime());
 
-        return calendarRepository.saveSchedule(calendar, writer);
+        return calendarRepository.saveSchedule(calendar, writer, writerId);
     }
 
     @Override
@@ -40,28 +47,59 @@ public class CalendarServiceImpl implements CalendarService {
             return calendarRepository.findAllSchedules();
         }
 
-        return calendarRepository.findAllSchedules(dto.getModDate(), dto.getWriterName(), dto.getWriterId(), dto.getEmail());
+        return calendarRepository.findAllSchedules(dto.getModDate(), dto.getWriterName(), dto.getWriterId(),
+                dto.getEmail());
     }
 
     @Override
     public CalendarResponseDto findScheduleById(Long id) {
 
-        return calendarRepository.findScheduleById(id);
+        return calendarRepository.findScheduleByIdOrElseThrow(id);
     }
 
     @Transactional
     @Override
     public CalendarResponseDto updateSchedule(Long id, CalendarRequestDto dto) {
 
-        calendarRepository.updateSchedule(id, dto.getTodo(), dto.getWriterName(), dto.getPassword(),
+        Long scheduleWriterId = calendarRepository.findWriterIdByScheduleId(id);
+        Long writerId = calendarRepository.findWriterIdByNameAndPassword(dto.getWriterName(), dto.getPassword());
+
+        if (writerId == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "작성자명 혹은 비밀번호가 틀렸습니다.");
+        }
+
+        if (!writerId.equals(scheduleWriterId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "본인의 일정만 수정할 수 있습니다.");
+        }
+
+        int updatedRow = calendarRepository.updateSchedule(id, dto.getTodo(),
                 Timestamp.valueOf(LocalDateTime.now()).toLocalDateTime());
 
-        return calendarRepository.findScheduleById(id);
+        if (updatedRow == 0) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 일정이 존재하지 않습니다.");
+        }
+
+        return calendarRepository.findScheduleByIdOrElseThrow(id);
     }
 
     @Override
     public void deleteSchedule(Long id, DeleteCalendarRequestDto dto) {
 
-        calendarRepository.deleteSchedule(id, dto.getEmail(), dto.getPassword());
+        Long scheduleWriterId = calendarRepository.findWriterIdByScheduleId(id);
+        Long writerId = calendarRepository.findWriterIdByEmailAndPassword(dto.getEmail(), dto.getPassword());
+
+        if (writerId == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "작성자명 혹은 비밀번호가 틀렸습니다.");
+        }
+
+        if (!writerId.equals(scheduleWriterId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "본인의 일정만 수정할 수 있습니다.");
+        }
+
+        int updatedRow = calendarRepository.deleteSchedule(id, dto.getEmail(), dto.getPassword());
+
+        if (updatedRow == 0) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 일정이 존재하지 않습니다.");
+        }
     }
 }
